整合背景图像js
